moved relevant test cases to test_solvers.py
